fixes to many things from the PR
