+++ .gitignore
@@ -9,4 +9,5 @@ node_modules
 .DS_Store
 dist/
 node_modules/
+.env
