+++ app.ts
@@ -1,6 +1,8 @@
 import { Application, Request, Response } from "express";
-import { env } from "process";
 import config from "./config"
+import * as dotenv from 'dotenv';
+
+dotenv.config();
 
 const express = require("express");
 const path = require("path");
@@ -8,6 +10,12 @@ const nunjucks = require("nunjucks");
 const app = express();
 const session = require("express-session");
 const bodyParser = require("body-parser");
+
+declare module "express-session" {
+    interface SessionData {
+        isLoggedIn: boolean;
+    }
+}
 
 // Nunjucks Configuration
 const appViews = path.join(__dirname, "/views/");
@@ -45,6 +53,12 @@ declare module "express-session" {
 
 app.listen(3000, () => {
     console.log("Server running on port 3000");
+});
+
+app.get("/", async (req: Request, res: Response) => {
+    const isLoggedIn: boolean = req.session.isLoggedIn;
+
+    res.render("index", { isLoggedIn, title: "Home" });
 });
 
 require("./controller/jobController")(app);
