+++ controller/authController.ts
@@ -1,21 +1,14 @@
 import { Request, Response, Application } from "express";
 import { Login } from "../model/auth";
-import { SessionData } from "express-session";
 
 const authService = require("../service/authService");
 const session = require("express-session");
 
-interface CustomSessionData extends SessionData {
-    isLoggedIn: boolean;
-}
-
-declare module "express-session" {
-    interface SessionData {
-        isLoggedIn: boolean;
-    }
-}
-
 module.exports = function(app: Application) {
+    app.get("/login", async (req: Request, res: Response) => {
+        res.render("login");
+    });
+
     app.post("/login", async (req: Request, res: Response) => {
         const data: Login = req.body;
