+++ package.json
@@ -12,11 +12,13 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/dotenv": "^8.2.0",
     "@types/express-session": "^1.17.10",
     "axios": "^1.6.2",
     "body-parser": "^1.20.2",
     "bootstrap": "^5.3.2",
     "chromedriver": "^120.0.0",
+    "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "express-session": "^1.17.3",
     "nunjucks": "^3.2.4",
