+++ service/authService.ts
@@ -1,9 +1,14 @@
+import * as dotenv from "dotenv";
 import { Login } from "../model/auth";
 const axios = require("axios");
 
+dotenv.config();
+
+const { API_URL } = process.env;
+
 module.exports.login = async function (login: Login): Promise<void> {
     try {
-        const response = await axios.post("http://localhost:8080/api/login", login);
+        const response = await axios.post(`${API_URL}/api/login`, login);
 
         return response.data;
     } catch (e) {
