+++ service/jobService.ts
@@ -1,19 +1,25 @@
+import * as dotenv from "dotenv";
 import { Job } from "../model/job";
-const axios = require('axios');
+const axios = require("axios");
+
+dotenv.config();
+
+const { API_URL } = process.env;
+
+console.log(API_URL)
 
 module.exports.getJobs = async function (): Promise<Job> {
     try {
-        const response = await axios.get("http://localhost:8080/api/job-roles");
+        const response = await axios.get(`${API_URL}/api/job-roles`);
         return response.data;
     } catch (e) {
         throw new Error("Could not get job");
     }
 };
 
-
 module.exports.getJobSpecById = async function (id: number): Promise<Job> {
     try {
-        const response = await axios.get("http://localhost:8080/api/job-specification/" + id);
+        const response = await axios.get(`${API_URL}/api/job-specification/` + id);
         return response.data;
     } catch (e) {
         throw new Error("Could not get job specification by id");
