+++ views/index.html
@@ -28,14 +28,6 @@
                 </ul>
             </div>
             {% else %}
-            <div class="navbar-collapse" id="navbarNav">
-                <ul class="navbar-nav mr-auto">
-                    <li class="nav-item ml-4">
-                    </li>
-                    <li class="nav-item ml-2">
-                    </li>
-                </ul>
-            </div>
             {% endif %}
             <div class="navbar-collapse justify-content-end mr-4">
             {% if isLoggedIn %}
