+++ views/login.html
@@ -34,10 +34,10 @@
                               </div>
                               <form action="/login" method="post">
                                 <div class="form-outline form-white mb-5">
-                                  <input type="email" id="username" name="username" class="form-control form-control-lg shadow-sm" placeholder="Email" value="{{ username }}" />
+                                  <input type="email" id="username" name="username" class="form-control form-control-lg shadow-sm" placeholder="Email" value="{{ username }}"/>
                                 </div>
                                 <div class="form-outline form-white mt-5 mb-5">
-                                  <input type="password" id="password" name="password" class="form-control form-control-lg shadow-sm" placeholder="Password" value="{{ password }}" />
+                                  <input type="password" id="password" name="password" class="form-control form-control-lg shadow-sm" placeholder="Password"/>
                                 </div>              
                                 <button id="btn-login" class="btn-login btn-lg px-5 shadow-sm" type="submit">Login</button>
                               </form>
